# apps/cloud-backend/fusion-core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.AuthTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 86400
        type: integer
      id_token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.PriceDetail:
    properties:
      currency:
        type: string
      price:
        type: number
      variant:
        type: string
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.PriceResponse:
    properties:
      prices:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.PriceDetail'
        type: array
      product_id:
        type: integer
      version:
        type: string
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse:
    properties:
      assets: {}
      description:
        type: string
      is_fusion_compatible:
        type: boolean
      model_family:
        type: string
      model_name:
        type: string
      product_id:
        type: integer
      specifications: {}
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductResponse:
    properties:
      accessory:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
        type: array
      amplifier:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
        type: array
      controller:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
        type: array
      dsp:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
        type: array
      io_endpoint:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
        type: array
      speaker:
        items:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
        type: array
      version:
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.SingleProductResponse:
    properties:
      accessory:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
      amplifier:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
      controller:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
      dsp:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
      io_endpoint:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
      speaker:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductItemResponse'
      version:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Fusion Cloud Backend API server providing comprehensive
    role-based access control, user management, and project management capabilities.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Fusion Cloud Backend API
  version: "1.0"
paths:
  /auth/automation/tokens:
    get:
      consumes:
      - application/json
      description: Get access and ID tokens for a user using Resource Owner Password
        flow (QA/Staging environment only)
      parameters:
      - description: Username for token generation
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tokens generated successfully
          schema:
            $ref: '#/definitions/types.AuthTokenResponse'
        "400":
          description: Bad request - username is required
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: auth automation endpoint is disabled
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Auth0 tokens for automation testing
      tags:
      - devices
...
  /products:
    get:
      consumes:
      - application/json
      description: Get all available products including speakers, amplifiers, DSPs,
        controllers, and endpoints
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with all products
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProductResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get all products
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific product by its unique identifier
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with product details
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.SingleProductResponse'
        "400":
          description: Bad request - Product ID is required
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get product by ID
      tags:
      - products
  /products/{id}/prices:
    get:
      consumes:
      - application/json
      description: Get prices for a specific product by its unique identifier, optionally
        filtered by currency
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Currency code (e.g., USD, EUR) to filter prices
        in: query
        name: currency
        type: string
      - description: Product variant to filter prices
        in: query
        name: variant
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with product prices
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.PriceResponse'
        "400":
          description: Bad request - Product ID is required
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get product prices by ID
      tags:
      - products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"